Refactored BigCube and SmallCube controllers into one script
