--- Assets/Scripts/Controllers/SmallCubeController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 72db85547e22d2b46a55957a0de836af
-timeCreated: 1437585550
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
