+++ Assets/Scripts/libraries.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d27dc6f4ba83d8a4b9e4b3829dd61a74
+folderAsset: yes
+timeCreated: 1439916469
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
