+++ Assets/UnityVS/Editor/UnityVS.VersionSpecific.dll.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: 4d490df8d9b6df848ba3c26594c7fa66
-timeCreated: 1440944761
+timeCreated: 1441568050
 licenseType: Free
 PluginImporter:
   serializedVersion: 1
